welcomehome/requirement.txt
black==24.10.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
